added some plots and error results (*ASGS*) for this version
I also made a few changes that only effect when we use ('velocity'). To complete that version we really need to add the gradients or divergence  of the ('velocity') array, but in the pwl-bdm post-processing the divergence is point-wise zero so I just set the velocity gradients to zero for now.